Remove duplicate systemd unit file for MongoDB
The Mongo role previously installed a unit file for MongoDB even though
the package includes a more robust version.  The service was started in
such a way that the package's unit file is used to run the service
initially, but the extra unit file is used to start the service upon
rebooting.  This could lead to an inconsistent experience, so this
change just uses the package-provided unit file for everything.